requirement.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.3
python-dotenv==1.0.0
tenacity==8.2.3
lxml==4.9.3
playwright==1.40.0